requirements.txt
certifi==2019.11.28
Django==1.11
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.9.4
pytz==2019.3
